+++ apps/web/src/common/routes/TermsRoute.tsx
@@ -43,6 +43,7 @@ export function TermsRoute() {
     control,
     handleSubmit,
     formState: { errors },
+    watch,
   } = useForm<TermsFormValues>({
     resolver: zodResolver(schema),
     defaultValues: {
@@ -55,10 +56,12 @@ export function TermsRoute() {
   if (!currentTerms) return null
 
   const onSubmit = ({ aiUsagePolicyAccepted }: TermsFormValues) => {
-    dispatch(acceptTerms({ aiUsagePolicyAccepted }))
+    dispatch(acceptTerms({ aiUsagePolicyAccepted, onSuccess: () => window.location.reload() }))
   }
 
   const showMandatoryError = !!(errors.generalConditionsAccepted || errors.privacyPolicyAccepted)
+
+  const disabled = !(watch("generalConditionsAccepted") && watch("privacyPolicyAccepted"))
 
   return (
     <FullPageCenterLayout className="min-h-screen p-4">
@@ -115,7 +118,9 @@ export function TermsRoute() {
               </p>
             )}
 
-            <Button type="submit">{t("termsAcceptance:submit")}</Button>
+            <Button disabled={disabled} type="submit">
+              {t("termsAcceptance:submit")}
+            </Button>
           </form>
         </CardContent>
       </Card>